requirements.txt
discord.py==2.3.2
deepl==1.18.0
